+++ hpx/parallel/segmented_algorithms/for_each.hpp
@@ -16,8 +16,8 @@
 #include <hpx/parallel/algorithms/detail/algorithm_result.hpp>
 #include <hpx/parallel/algorithms/detail/dispatch.hpp>
 #include <hpx/parallel/algorithms/detail/is_negative.hpp>
-#include <hpx/parallel/algorithms/remote/dispatch.hpp>
 #include <hpx/parallel/algorithms/for_each.hpp>
+#include <hpx/parallel/segmented_algorithms/detail/dispatch.hpp>
 #include <hpx/parallel/util/detail/handle_remote_exceptions.hpp>
 
 #include <algorithm>
@@ -59,7 +59,7 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 local_iterator_type end = traits::local(last);
                 if (beg != end)
                 {
-                    util::remote::dispatch(traits::get_id(sit),
+                    dispatch(traits::get_id(sit),
                         std::forward<Algo>(algo), policy, true_(),
                         beg, end, std::forward<F>(f));
                 }
@@ -70,7 +70,7 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 local_iterator_type end = traits::end(sit);
                 if (beg != end)
                 {
-                    util::remote::dispatch(traits::get_id(sit),
+                    dispatch(traits::get_id(sit),
                         std::forward<Algo>(algo), policy, true_(),
                         beg, end, std::forward<F>(f));
                 }
@@ -82,7 +82,7 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                     end = traits::end(sit);
                     if (beg != end)
                     {
-                        util::remote::dispatch(traits::get_id(sit),
+                        dispatch(traits::get_id(sit),
                             std::forward<Algo>(algo), policy, true_(),
                             beg, end, std::forward<F>(f));
                     }
@@ -93,7 +93,7 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 end = traits::local(last);
                 if (beg != end)
                 {
-                    util::remote::dispatch(traits::get_id(sit),
+                    dispatch(traits::get_id(sit),
                         std::forward<Algo>(algo), policy, true_(),
                         beg, end, std::forward<F>(f));
                 }
@@ -132,10 +132,9 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 local_iterator_type end = traits::local(last);
                 if (beg != end)
                 {
-                    segments.push_back(
-                        util::remote::dispatch_async(traits::get_id(sit),
-                            std::forward<Algo>(algo), policy, forced_seq(),
-                            beg, end, std::forward<F>(f))
+                    segments.push_back(dispatch_async(traits::get_id(sit),
+                        std::forward<Algo>(algo), policy, forced_seq(),
+                        beg, end, std::forward<F>(f))
                     );
                 }
             }
@@ -145,10 +144,9 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 local_iterator_type end = traits::end(sit);
                 if (beg != end)
                 {
-                    segments.push_back(
-                        util::remote::dispatch_async(traits::get_id(sit),
-                            std::forward<Algo>(algo), policy, forced_seq(),
-                            beg, end, std::forward<F>(f))
+                    segments.push_back(dispatch_async(traits::get_id(sit),
+                        std::forward<Algo>(algo), policy, forced_seq(),
+                        beg, end, std::forward<F>(f))
                     );
                 }
 
@@ -159,10 +157,9 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                     end = traits::end(sit);
                     if (beg != end)
                     {
-                        segments.push_back(
-                            util::remote::dispatch_async(traits::get_id(sit),
-                                std::forward<Algo>(algo), policy, forced_seq(),
-                                beg, end, std::forward<F>(f))
+                        segments.push_back(dispatch_async(traits::get_id(sit),
+                            std::forward<Algo>(algo), policy, forced_seq(),
+                            beg, end, std::forward<F>(f))
                         );
                     }
                 }
@@ -172,10 +169,9 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 end = traits::local(last);
                 if (beg != end)
                 {
-                    segments.push_back(
-                        util::remote::dispatch_async(traits::get_id(sit),
-                            std::forward<Algo>(algo), policy, forced_seq(),
-                            beg, end, std::forward<F>(f))
+                    segments.push_back(dispatch_async(traits::get_id(sit),
+                        std::forward<Algo>(algo), policy, forced_seq(),
+                        beg, end, std::forward<F>(f))
                     );
                 }
             }
