+++ hpx/parallel/segmented_algorithms/transform_reduce.hpp
@@ -16,8 +16,8 @@
 #include <hpx/parallel/algorithms/detail/algorithm_result.hpp>
 #include <hpx/parallel/algorithms/detail/dispatch.hpp>
 #include <hpx/parallel/algorithms/detail/is_negative.hpp>
-#include <hpx/parallel/algorithms/remote/dispatch.hpp>
 #include <hpx/parallel/algorithms/transform_reduce.hpp>
+#include <hpx/parallel/segmented_algorithms/detail/dispatch.hpp>
 #include <hpx/parallel/util/detail/handle_remote_exceptions.hpp>
 
 #include <algorithm>
@@ -63,7 +63,7 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 if (beg != end)
                 {
                     overall_result =
-                        util::remote::dispatch(traits::get_id(sit),
+                        dispatch(traits::get_id(sit),
                             std::forward<Algo>(algo), policy, true_(),
                             beg, end, std::forward<T>(init),
                             std::forward<Reduce>(red_op),
@@ -78,7 +78,7 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 {
                     overall_result = red_op(
                         overall_result,
-                        util::remote::dispatch(traits::get_id(sit),
+                        dispatch(traits::get_id(sit),
                             std::forward<Algo>(algo), policy, true_(),
                             beg, end, std::forward<T>(init),
                             std::forward<Reduce>(red_op),
@@ -95,7 +95,7 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                     {
                         overall_result = red_op(
                             overall_result,
-                            util::remote::dispatch(traits::get_id(sit),
+                            dispatch(traits::get_id(sit),
                                 std::forward<Algo>(algo), policy, true_(),
                                 beg, end, std::forward<T>(init),
                                 std::forward<Reduce>(red_op),
@@ -111,7 +111,7 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 {
                     overall_result = red_op(
                         overall_result,
-                        util::remote::dispatch(traits::get_id(sit),
+                        dispatch(traits::get_id(sit),
                             std::forward<Algo>(algo), policy, true_(),
                             beg, end, std::forward<T>(init),
                             std::forward<Reduce>(red_op),
@@ -155,12 +155,11 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 local_iterator_type end = traits::local(last);
                 if (beg != end)
                 {
-                    segments.push_back(
-                        util::remote::dispatch_async(traits::get_id(sit),
-                            std::forward<Algo>(algo), policy, forced_seq(),
-                            beg, end, std::forward<T>(init),
-                            std::forward<Reduce>(red_op),
-                            std::forward<Convert>(conv_op))
+                    segments.push_back(dispatch_async(traits::get_id(sit),
+                        std::forward<Algo>(algo), policy, forced_seq(),
+                        beg, end, std::forward<T>(init),
+                        std::forward<Reduce>(red_op),
+                        std::forward<Convert>(conv_op))
                     );
                 }
             }
@@ -170,12 +169,11 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 local_iterator_type end = traits::end(sit);
                 if (beg != end)
                 {
-                    segments.push_back(
-                        util::remote::dispatch_async(traits::get_id(sit),
-                            std::forward<Algo>(algo), policy, forced_seq(),
-                            beg, end, std::forward<T>(init),
-                            std::forward<Reduce>(red_op),
-                            std::forward<Convert>(conv_op))
+                    segments.push_back(dispatch_async(traits::get_id(sit),
+                        std::forward<Algo>(algo), policy, forced_seq(),
+                        beg, end, std::forward<T>(init),
+                        std::forward<Reduce>(red_op),
+                        std::forward<Convert>(conv_op))
                     );
                 }
 
@@ -186,12 +184,11 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                     end = traits::end(sit);
                     if (beg != end)
                     {
-                        segments.push_back(
-                            util::remote::dispatch_async(traits::get_id(sit),
-                                std::forward<Algo>(algo), policy, forced_seq(),
-                                beg, end, std::forward<T>(init),
-                                std::forward<Reduce>(red_op),
-                                std::forward<Convert>(conv_op))
+                        segments.push_back(dispatch_async(traits::get_id(sit),
+                            std::forward<Algo>(algo), policy, forced_seq(),
+                            beg, end, std::forward<T>(init),
+                            std::forward<Reduce>(red_op),
+                            std::forward<Convert>(conv_op))
                         );
                     }
                 }
@@ -201,12 +198,11 @@ namespace hpx { namespace parallel { HPX_INLINE_NAMESPACE(v1)
                 end = traits::local(last);
                 if (beg != end)
                 {
-                    segments.push_back(
-                        util::remote::dispatch_async(traits::get_id(sit),
-                            std::forward<Algo>(algo), policy, forced_seq(),
-                            beg, end, std::forward<T>(init),
-                            std::forward<Reduce>(red_op),
-                            std::forward<Convert>(conv_op))
+                    segments.push_back(dispatch_async(traits::get_id(sit),
+                        std::forward<Algo>(algo), policy, forced_seq(),
+                        beg, end, std::forward<T>(init),
+                        std::forward<Reduce>(red_op),
+                        std::forward<Convert>(conv_op))
                     );
                 }
             }
